forza/wsl_requirements.txt
coverage==7.13.5
numpy==2.4.3
pandas==3.0.1
python-afl==0.7.3
python-dateutil==2.9.0.post0
six==1.17.0
